updated form names to fit conventions, organized app.jsx, ensured that property nav is working as expected
